разделил на модули
так же исправил импорты, сделал понятнее функцию, теперь изменен ввод параметров в консоль, путь_до_папки путь_до_аннотации диапазоны размеров
